faire les controllers et les routes

+++ backend/controllers/sauce.js
@@ -1,6 +1,8 @@
 //ce fichier contient la logique de métier
 const Sauce = require('../models/sauce');
+const fs = require('fs');
 exports.creatSauce = (req, res, next) => {
+    const sauceObject = JSON.parse(req.body.sauce);
     //supprimé en amont le faux_id envoyé par le front-end
     //on suprime id parceque va etre generer pas BDD
     delete sauceObject._id;
@@ -11,7 +13,7 @@ exports.creatSauce = (req, res, next) => {
         //de tous les éléments de req.body
         ...sauceObject,
         userId: req.auth.userId,
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
     });
     // méthode save() qui enregistre simplement votre Thing dans la base de données.
     sauce.save()
@@ -27,7 +29,7 @@ exports.getOneSauce = (req, res, next) => {
         .catch(error => res.status(404).json({ error: error }));
 };
 
-exports.getAllSauce = (req, res, next) => {
+exports.getAllSauces = (req, res, next) => {
     // find() pour renvoyer un tableau contenant tous les sauces dans notre base de données
     Sauce.find()
         .then(sauces => res.status(200).json(sauces))
@@ -52,7 +54,7 @@ exports.modifySauce = (req, res, next) => {
                 res.status(401).json({ message: 'Not authorized' });
             } else {
                 Sauce.updateOne({ _id: req.params.id }, {...sauceObject, _id: req.params.id })
-                    .then(() => res.status(200).json({ message: 'sauce modifié!' }))
+                    .then(() => res.status(200).json({ message: 'Sauce modifié!' }))
                     .catch(error => res.status(401).json({ error }));
             }
         })
@@ -62,7 +64,20 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Sauce supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'Sauce supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
